Remove header extension
