+++ sdk/storage/src/core/mod.rs
@@ -18,6 +18,7 @@ pub use copy_progress::CopyProgress;
 pub(crate) mod parsing_xml;
 mod stored_access_policy;
 pub use errors::{Error, Result};
+pub(crate) mod util;
 pub mod xml;
 
 #[derive(Debug, Clone, Eq, PartialEq, Copy, Serialize, Deserialize)]
