+++ sdk/storage/src/table/responses/insert_entity_response.rs
@@ -1,8 +1,8 @@
+use crate::core::util::HeaderMapExt;
 use crate::EntityWithMetadata;
 use azure_core::{
     headers::{etag_from_headers, get_str_from_headers, CommonStorageResponseHeaders},
     prelude::Etag,
-    util::HeaderMapExt,
 };
 use bytes::Bytes;
 use http::Response;
